refactor: exports, types, hashing, body types, metrics logic
